Ailly plugin has correct internal paths and is .mjs
